requirements.txt
Flask==3.1.1
gunicorn==23.0.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Pillow==11.0.0
Flask-WTF==1.2.2
Werkzeug==3.1.3
requests==2.32.3
